+++ lib/ember_cli/missing_manifest.rb
@@ -1,0 +1,11 @@
+module EmberCli
+  class MissingManifest
+    def files
+      {}
+    end
+
+    def assets
+      {}
+    end
+  end
+end